+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * CWL: Fix for making target step argument for cwltool.
 * CWL: Added support for "metadata" field for input files in CWL under development (pending inclusion in CWL standard).
 * CWL: fix casting cwl types from scatter to scatter.
+* CWL: Fix for parameter mapping for output stage of the workflow.
 
 
 ## 2.10.2 2022-05-17